Move views hierarchy setup to new ContentView
